refactor: criar components dos elementos da home e profile para melhorar organização e reutilização
