+++ finance-dashboard/src/components/Header/Profile/index.tsx
@@ -1,0 +1,23 @@
+import {
+  ProfileHeaderContainer,
+  ProfileImage,
+  ProfileSection,
+  ProfileUserName,
+} from "./styled";
+
+const userImage = require("../../../../assets/user.png");
+
+interface HeaderProps {
+  name: string;
+}
+
+export function ProfileHeader({ name }: HeaderProps) {
+  return (
+    <ProfileHeaderContainer>
+      <ProfileSection>
+        <ProfileImage source={userImage} />
+        <ProfileUserName>{name}</ProfileUserName>
+      </ProfileSection>
+    </ProfileHeaderContainer>
+  );
+}
